Change taskManager to TaskAssigneeManager
